+++ lib/src/toast_alignment.dart
@@ -1,0 +1,12 @@
+/// Denotes the alignment of the toast on the screen.
+enum ToastAlignment {
+  top,
+  bottom,
+  center,
+  topLeft,
+  topRight,
+  bottomLeft,
+  bottomRight,
+  centerLeft,
+  centerRight,
+}