+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0
+
+- Added `ToastAnimation` class and some prebuilt implementations
+- Changed toast wrapping structure to use this new `ToastAnimation`
+
 ## 0.2.1
 
 - Changed the way the package retrieves the overlay internally to be more robust.